requirements.txt
appdirs==1.4.3
billiard==3.5.0.2
cffi==1.9.1
click==6.6
coreapi==2.1.1
coreapi-cli==1.0.6
coverage==4.2
Django==1.10.4
django-cors-headers==1.3.1
django-debug-toolbar==1.6
django-extensions==1.7.5
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-gis==0.11
djangorestframework-jwt==1.9.0
docutils==0.12
factory-boy==2.7.0
fake-factory==0.6.0
flake8==3.3.0
idna==2.1
isort==4.2.5
itypes==1.1.0
jdcal==1.3
jmespath==0.9.0
jws==0.1.3
kombu==4.0.0
Markdown==2.6.6
mccabe==0.6.1
oauth2client==3.0.0
openapi-codec==1.2.1
openpyxl==2.4.7
packaging==16.8
Pillow==3.3.1
protobuf==3.1.0.post1
pyaml==16.9.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycparser==2.17
pycryptodome==3.4.3
pydot==1.2.3
pyflakes==1.5.0
pyparsing==2.1.10
pytz==2016.10
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
vine==1.1.3
wheel==0.24.0
